--- a/types/Cargo.toml
+++ b/types/Cargo.toml
@@ -36,6 +36,8 @@ libra-time = { path = "../common/time", version = "0.1.0" }
 libra-workspace-hack = { path = "../common/workspace-hack", version = "0.1.0" }
 move-core-types = { path = "../language/move-core/types", version = "0.1.0" }
 
+unroll = "0.1.5"
+
 [dev-dependencies]
 regex = "1.3.9"
 proptest = "0.10.1"