Not supposed to be committed.
This PR will switch-off various presumably expensive components (e.g. signature verifications, SHA3 hashes, etc.) and run performance test to compare the TPS.

Results so far (for default 4500 accounts):
[1] all up (base-line): 1010 TPS
[2] all up without signature verifications: 1035 TPS (2.5% rise)
[3] all up with signature verifications done twice (i.e. slowed down by 2x): 904 TPS (10.5% drop)
[4] all up with verifications done 5 times (slowed down by 5x): 549 TPS (54% drop)

[5] all up without hashing of internal Merkle Nodes: 1125 TPS (10% rise)
he CryptoHash for MerkleTreeInternalNode is made to be:
```
        let mut result = [0u8; HashValue::LENGTH];
        for i in 0..HashValue::LENGTH {
            result[i] = self.left_child[i] ^ self.right_child[i];
        }
        HashValue::new(result)
```
[6] all up with hashing of internal Merkle Nodes done twice, which simulates a tree with 20 million accounts (4500^2), tree of height 24: 904 TPS (10% drop)
[7] all up with hashing of internal Merkle Nodes repeated three times (simulates 4500^2 = 91 billion accounts, tree of height 36): 824 TPS (18% drop), 23% raise in latency

Full output:
```
[1]
all up : 1010 TPS, 4475 ms latency, 5400 ms p99 latency, no expired txns
10% down : 995 TPS, 4092 ms latency, 5000 ms p99 latency, no expired txns
3 Region Simulation : 756 TPS, 6002 ms latency, 7750 ms p99 latency, no expired txns
fixed tps 10 : 10 TPS, 603 ms latency, 800 ms p99 latency, no expired txns
```
```
[2]
all up : 1035 TPS, 4373 ms latency, 5300 ms p99 latency, no expired txns
10% down : 987 TPS, 4113 ms latency, 6600 ms p99 latency, no expired txns
3 Region Simulation : 781 TPS, 5799 ms latency, 7700 ms p99 latency, no expired txns
fixed tps 10 : 10 TPS, 577 ms latency, 600 ms p99 latency, no expired txns
```
```
[3]
all up : 904 TPS, 5015 ms latency, 5900 ms p99 latency, no expired txns
10% down : 883 TPS, 4621 ms latency, 7600 ms p99 latency, no expired txns
3 Region Simulation : 709 TPS, 6386 ms latency, 8500 ms p99 latency, no expired txns
fixed tps 10 : 10 TPS, 712 ms latency, 800 ms p99 latency, no expired txns
```
```
[4]
all up : 549 TPS, 8280 ms latency, 11150 ms p99 latency, no expired txns
```
```
[5]
all up : 1125 TPS, 4023 ms latency, 4850 ms p99 latency, no expired txns
```
```
[6]
all up : 904 TPS, 5017 ms latency, 6050 ms p99 latency, no expired txns
```
```
[7] 824 TPS, 5509 ms latency, 6950 ms p99 latency, no expired txns
```